Small update to help newbies a bit.
A common error with tables is that you miss a comma whilst defining them. For those who know Lua, searching for missing commas on an error like this is just a "standard procedure." However, for those who are new to Lua, they'll see the error and likely take it at face-value, looking instead for some spot they may have missed a closing bracket.

![Reasoning: a table missing a comma outputting an error stating it is missing a closing bracket.](https://media.discordapp.net/attachments/872883444970033243/1076191867512819802/SmartSelect_20230217_102205_Discord.jpg)

# Quick notice
I don't know if this module has any tests or if this will affect anything else negatively. I just had the idea today when looking at an error someone got on the CC:Aide bot and decided to quickly throw in a PR from my phone while at work.