Build the web-emulator as part of the doc site
Historically we've used copy-cat to provide a web-based emulator to run
example code on our doc site. However, copy-cat is often out-of-date
with CC:T, which means example snippets fail when you try to run them!

This copies copy-cat into the main project, allowing us to ensure the
two are always in sync.